+++ t/099-mk-sandbox.t
@@ -16,9 +16,20 @@ my MongoDB::Test-support $ts .= new;
 
 #-------------------------------------------------------------------------------
 for $ts.server-range -> $server-number {
-  ok $ts.server-control.start-mongod("s$server-number"),
-     "Server $server-number started";
+  try {
+    ok $ts.server-control.start-mongod("s$server-number"),
+       "Server $server-number started";
+    CATCH {
+      when X::MongoDB::Message {
+        like .message, /:s exited unsuccessfully /,
+             "Server 's$server-number' already started";
+      }
+    }
+  }
 }
+
+throws-like { $ts.server-control.start-mongod('s1') },
+            X::MongoDB::Message, :message(/:s exited unsuccessfully/);
 
 #-------------------------------------------------------------------------------
 # Cleanup and close