+++ xt/Sxml/driver-authentication.t
@@ -1,4 +1,4 @@
-use v6.c;
+use v6;
 use Test;
 
 use lib 't';